requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
cachelib==0.9.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Flask==2.2.2
Flask-Celery-Helper==1.1.0
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-Session==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
packaging==21.3
prettytable==3.5.0
prompt-toolkit==3.0.33
protobuf==3.20.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.6
redis==4.3.5
rq==1.11.1
six==1.16.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
